Move to ecCodes
The decision has been made to move from supporting both the stand-alone
GRIB API and ecCodes to solely ecCodes. It did not seem prudent to keep
both options around when ECMWF fully intends to do all new development
on ecCodes. The major issue for users will be having to link to a new
library that is not as readily available on package managers. The rest
of the changes are internal and should not affect users. Internal name
changes to the functions are nearly ubiquitous. A few private functions
still go with the old libraries prefix. Will need to monitor for any
changes. Other changes include changing the environmental variables
users use to build the package from source. Per 'Writing R Extensions'
have changes PKG_CFLAGS to PKG_CPPFLAGS and PKG_LDFLAGS to PKG_LIBS.