Fix file descriptor leak; minor cleanup
Main issue addressed in this commit is to fix the file descriptor leak
that existed in the GRIB file test of the file open process. This led
to users with large amounts of files to process to hit the per-process,
open file descriptor limit imposed by *nix systems. The file was
opened during the testing before the actual file handle was sent to R
and then never closed. Simply closig the file after the test alleviated
the problem.

Secondary to the descriptor leak, minor code cleaup occurred to update
the version and contact information. BugReports was added to the
DESCRIPTION in order to nudge users to that venue for bug reporting.

Finally, some travis testing environments were added to make sure the
package is tested against more versions of ecCodes.

Fixes #12