Fix grib_select file descriptor leak
grib_select was implemented with a separate fopen call in order to get
a codes_handle to use for getting the native types. The fopen did not
have a corresponding fclose. This solution removes the fopen call and
uses the available file pointer from the grib_open call. The file path
is still needed to add to the codes_index, but the fopen/fclose done in
those functions should not conflict with the pointer from grib_open.

This also does some housekeeping with the CI/coverage config as things
have changed over time. Download link for ECCODES was updated. Docs now
built with latest roxygen2. Package building scripts were updated.

Related to #16