Updated docs/links
Primary changes are to update the links to the ECMWF documentation as
many links no longer go to working pages. Removed references to the
GRIB API as that has not been in development for some time. Made a
change to the environment variable used to point to the GRIB
defintion files. This would only affect Windows as the same change
was already implemented in ECCODES. With Windows still not working,
it is not likely this change will impact many users.