Cleanup before v1.2.3 release
Cleaned up old references in the docs and updated dates/versions where
they were stale. Removed grib_context_full_defs_path as there is now
a codes_definition_path function that does not require accessing a
private function. Tests indicate this should work the same.

ecCodes >= 2.19.0 is now a requirement. There was an bug in the way
ecCodes deleted grib_index objects (see issue ECC-440). With this
impacting users in certain cases of using grib_select on multiple files
it was the best move to make.