Fix grib_select file descriptor leak (#17)
* Fix grib_select file descriptor leak

grib_select was implemented with a separate fopen call in order to get
a codes_handle to use for getting the native types. The fopen did not
have a corresponding fclose. This solution removes the fopen call and
uses the available file pointer from the grib_open call. The file path
is still needed to add to the codes_index, but the fopen/fclose done in
those functions should not conflict with the pointer from grib_open.

This also does some housekeeping with the CI/coverage config as things
have changed over time. Download link for ECCODES was updated. Docs now
built with latest roxygen2. Package building scripts were updated.

Related to #16

* Updated docs/links

Primary changes are to update the links to the ECMWF documentation as
many links no longer go to working pages. Removed references to the
GRIB API as that has not been in development for some time. Made a
change to the environment variable used to point to the GRIB
defintion files. This would only affect Windows as the same change
was already implemented in ECCODES. With Windows still not working,
it is not likely this change will impact many users. Added note for change
in required ecCodes version.

* Cleanup before v1.2.3 release

Cleaned up old references in the docs and updated dates/versions where
they were stale. Removed grib_context_full_defs_path as there is now
a codes_definition_path function that does not require accessing a
private function. Tests indicate this should work the same.

ecCodes >= 2.19.0 is now a requirement. There was an bug in the way
ecCodes deleted grib_index objects (see issue ECC-440). With this
impacting users in certain cases of using grib_select on multiple files
it was the best move to make.